.jshintrc deletion
Hello, I wanted to contribute to code, but my IDE is having trouble with jshintrc rules (WebStorm, it automatically parses it and then syntax-higlights the problems). Are we using it anywhere? I couldn't find any usage of it, neither is jshint in package.json. I think the jshintrc is there as the project historically used jshint, but it does not use it anymore. Can jshintrc be removed?

(file is 'src/.jshintrc')